third_party/blink/web_tests/http/tests/security/mixedContent/insecure-formSubmission-in-main-frame-blocked-expected.txt

CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-formSubmission.html' was loaded over a secure connection, but contains a form that targets an insecure endpoint 'http://example.test:8080/security/resources/boring.html'. This endpoint should be made available over a secure connection.
This test opens a window that shows a form with "action" pointing to an insecure location. We should not trigger a mixed content callback even though the main frame in the window is HTTPS and the form is pointing to insecure content, because we've set the preference to block this.
